fix slider elements to respond properly to horizontal scroll
