+++ src/components/pages/DataVisualizations/YearLimitsSelect.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Form, Button, Input } from 'antd';
 import {
   setVisualizationData,
